Fix the action_call_table control plane.
Refactor the ebpf python infrastructure.
