+++ testdata/p4_16_samples/action_call_table_ebpf.stf
@@ -1,6 +1,6 @@
 # bit<32> A bit<32> B
 # In the output B = (A + 10)
-setdefault pipe_t pipe_Reject(rej:1)
+setdefault pipe_t pipe_Reject(rej:0)
 
 
 packet 0 00000000 ABCDEF01
